Enemy bullets are now cool. Also player has a trigger collider now that detecxts bullets, can read tha in the description or slack
Okay the bullets now face the night direction. Also I made it so the player now has a trigger collider as well, This is because the enemy bullets have to be triggers because they would get stuck on the enemies. Also it helps not that the player can get hit by bullets across their body rather than just the legs (but the normal collider is still just legs to give the "illusion of 3D")